pyproject.toml
[project]
name = "internet-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastparquet>=2024.11.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "ping3>=4.0.8",
    "plotly>=6.0.0",
    "pyarrow>=19.0.0",
    "seaborn>=0.13.2",
    "speedtest-cli>=2.1.3",
]
